Gutter data working. Added boilerplate code for autolog
