+++ src/traceService/models/auto-log-tracer.js
@@ -1,0 +1,171 @@
+export class AutoLogTracer{
+    constructor(){
+        
+    }
+    getAutologCodeBoilerPlate(){
+        return `
+        var Syntax = {
+        AssignmentExpression: 'AssignmentExpression',
+        ArrayExpression: 'ArrayExpression',
+        BlockStatement: 'BlockStatement',
+        BinaryExpression: 'BinaryExpression',
+        BreakStatement: 'BreakStatement',
+        CallExpression: 'CallExpression',
+        CatchClause: 'CatchClause',
+        ConditionalExpression: 'ConditionalExpression',
+        ContinueStatement: 'ContinueStatement',
+        DoWhileStatement: 'DoWhileStatement',
+        DebuggerStatement: 'DebuggerStatement',
+        EmptyStatement: 'EmptyStatement',
+        ExpressionStatement: 'ExpressionStatement',
+        ForStatement: 'ForStatement',
+        ForInStatement: 'ForInStatement',
+        FunctionDeclaration: 'FunctionDeclaration',
+        FunctionExpression: 'FunctionExpression',
+        Identifier: 'Identifier',
+        IfStatement: 'IfStatement',
+        Literal: 'Literal',
+        LabeledStatement: 'LabeledStatement',
+        LogicalExpression: 'LogicalExpression',
+        MemberExpression: 'MemberExpression',
+        NewExpression: 'NewExpression',
+        ObjectExpression: 'ObjectExpression',
+        Program: 'Program',
+        Property: 'Property',
+        ReturnStatement: 'ReturnStatement',
+        SequenceExpression: 'SequenceExpression',
+        SwitchStatement: 'SwitchStatement',
+        SwitchCase: 'SwitchCase',
+        ThisExpression: 'ThisExpression',
+        ThrowStatement: 'ThrowStatement',
+        TryStatement: 'TryStatement',
+        UnaryExpression: 'UnaryExpression',
+        UpdateExpression: 'UpdateExpression',
+        VariableDeclaration: 'VariableDeclaration',
+        VariableDeclarator: 'VariableDeclarator',
+        WhileStatement: 'WhileStatement',
+        WithStatement: 'WithStatement'
+    };
+    var traceTypes = {
+        LocalStack : [Syntax.FunctionDeclaration, Syntax.FunctionExpression],
+        Expression: [
+            Syntax.UnaryExpression,
+            Syntax.UpdateExpression,
+            Syntax.CallExpression,
+            Syntax.Property,
+            Syntax.VariableDeclarator,
+            Syntax.AssignmentExpression,
+            Syntax.BinaryExpression,
+            Syntax.ReturnStatement,
+            Syntax.ForStatement,
+            Syntax.ForInStatement,
+            Syntax.WhileStatement,
+            Syntax.DoWhileStatement,
+            Syntax.ExpressionStatement
+            ],
+        ExpressionStatement : [
+            Syntax.ExpressionStatement
+            ],
+        ControlFlow : [],
+        Condition: [],
+        Loop: [Syntax.WhileStatement],
+        exception: []
+        
+    };
+        window.CANTRACE = true;
+        window.ISCANCELLED = false;
+        window.TRACE = {
+            hits: {}, data: {}, stack : [], execution : [], variables: [], values : [], timeline: [], identifiers: [], 
+            autoLog: function autoLog(info) {
+                var key = info.text + ':' + info.indexRange[0]+':' + info.indexRange[1];
+                
+                if(traceTypes.LocalStack.indexOf(info.type)>-1){
+    				this.stack.push(key) ;
+                }
+
+                if(info.type === Syntax.VariableDeclarator || info.type === Syntax.AssignmentExpression){
+                   this.values.push({'id': info.id , 'value': JSON.stringify(info.value), 'range': info.range}); 
+                }
+
+                this.timeline.push({ id: info.id , value: JSON.stringify(info.value), range: info.range, type: info.type, text: info.text});
+
+
+                var stackTop =	this.stack.length - 1;
+                
+				if (this.hits.hasOwnProperty(key)) {
+                    this.hits[key] = this.hits[key] + 1;
+                    this.data[key].hits = this.hits[key] + 1;
+                    this.data[key].values.push({'stackIndex': stackTop, 'value' :JSON.stringify(info.value)});
+                } else {
+                    
+                    if(info.type === Syntax.VariableDeclarator){
+                       this.variables.push({'id': info.id , 'range': info.range});
+                    }
+                    
+                    this.identifiers.push({'id': info.id , 'range': info.range});
+                    
+                    
+                    this.hits[key] = 1;
+                    this.execution.push(key);
+                    this.data[key] = {
+                        'type' : info.type,
+                        'id' : info.id,
+                        'text' : info.text,
+                        'values': [{'stackIndex': stackTop, 'value' :JSON.stringify(info.value)}],
+                        'range': info.range,
+                        'hits' : 1,
+                        'extra' : info.extra
+                    };
+                }
+                
+                if(window.ISCANCELLED){
+                    throw "Trace Cancelled.";
+                }
+                
+                return info.value;
+            },
+            getStackTrace: function getStackTrace () {
+                var entry,
+                    stackData = [];
+                for (var i in this.stack) {
+                    if (this.stack.hasOwnProperty(i)) {
+                        entry = this.stack[i];
+                        stackData.push({ index: i, text: entry.split(':')[0], range: this.data[entry].range,  count: this.hits[entry]});
+                    }
+                }
+                return stackData;
+            },
+            getExecutionTraceAll: function getExecutionTraceAll() {
+                let result = [];
+                for (let i in this.execution) {
+                    let entry = this.execution[i];
+                    if (this.data.hasOwnProperty(entry)) {
+                        result.push(this.data[entry]);
+                    }
+                }
+                return result;
+            },
+            getExpressions: function getExpressions() {
+                return {variables : this.identifiers, timeline: this.timeline};
+            },
+            getVariables: function getVariables() {
+                return {variables : this.variables, values: this.values};
+            },
+            getExecutionTrace: function getExecutionTrace() {// getValues
+                var i, entry, data, stackData = [];
+                for (i in this.execution) {
+                    entry = this.execution[i];
+                    if (this.data.hasOwnProperty(entry)) {
+                        data =this.data[entry];
+                        if(traceTypes.Expression.indexOf(data.type) > -1  ){
+                            stackData.push(this.data[entry]);
+                        }
+                     }
+                }
+                return stackData;
+            }
+        };
+        
+        `;
+    }
+}
